fix some arguments
