+++ hydromt/models/model.py
@@ -7,7 +7,6 @@ import typing
 from abc import ABCMeta
 from inspect import _empty, signature
 from os.path import isabs, isfile, join
-from pathlib import Path
 from typing import (
     Any,
     Dict,
@@ -19,8 +18,6 @@ from typing import (
     cast,
 )
 
-import pandas as pd
-import xarray as xr
 from pyproj import CRS
 
 from hydromt import hydromt_step
@@ -447,162 +444,6 @@ class Model(object, metaclass=ABCMeta):
             is_equal &= component_equal
             errors.update(**component_errors)
         return is_equal, errors
-
-    # map files setup methods
-    def setup_maps_from_rasterdataset(
-        self,
-        raster_fn: Union[str, Path, xr.Dataset],
-        variables: Optional[List] = None,
-        fill_method: Optional[str] = None,
-        name: Optional[str] = None,
-        reproject_method: Optional[str] = None,
-        split_dataset: Optional[bool] = True,
-        rename: Optional[Dict] = None,
-    ) -> List[str]:
-        """HYDROMT CORE METHOD: Add data variable(s) from ``raster_fn`` to maps object.
-
-        If raster is a dataset, all variables will be added unless ``variables``
-        list is specified.
-
-        Adds model layers:
-
-        * **raster.name** maps: data from raster_fn
-
-        Parameters
-        ----------
-        raster_fn: str, Path, xr.Dataset
-            Data catalog key, path to raster file or raster xarray data object.
-        variables: list, optional
-            List of variables to add to maps from raster_fn. By default all.
-        fill_method : str, optional
-            If specified, fills nodata values using fill_nodata method.
-            Available methods are {'linear', 'nearest', 'cubic', 'rio_idw'}.
-        name: str, optional
-            Name of new dataset in self.maps dictionnary,
-            only in case split_dataset=False.
-        reproject_method: str, optional
-            See rasterio.warp.reproject for existing methods, by default the data is
-            not reprojected (None).
-        split_dataset: bool, optional
-            If data is a xarray.Dataset split it into several xarray.DataArrays.
-        rename: dict, optional
-            Dictionary to rename variable names in raster_fn before adding to maps
-            {'name_in_raster_fn': 'name_in_maps'}. By default empty.
-
-        Returns
-        -------
-        list
-            Names of added model map layers
-        """
-        rename = rename or {}
-        self.logger.info(f"Preparing maps data from raster source {raster_fn}")
-        # Read raster data and select variables
-        ds = self.data_catalog.get_rasterdataset(
-            raster_fn,
-            geom=self.region,
-            buffer=2,
-            variables=variables,
-            single_var_as_array=False,
-        )
-        # Fill nodata
-        if fill_method is not None:
-            ds = ds.raster.interpolate_na(method=fill_method)
-        # Reprojection
-        if ds.rio.crs != self.crs and reproject_method is not None:
-            ds = ds.raster.reproject(dst_crs=self.crs, method=reproject_method)
-        # Rename and add to maps
-        self.set_maps(ds.rename(rename), name=name, split_dataset=split_dataset)
-
-        return list(ds.data_vars.keys())
-
-    def setup_maps_from_raster_reclass(
-        self,
-        raster_fn: Union[str, Path, xr.DataArray],
-        reclass_table_fn: Union[str, Path, pd.DataFrame],
-        reclass_variables: List,
-        variable: Optional[str] = None,
-        fill_method: Optional[str] = None,
-        reproject_method: Optional[str] = None,
-        name: Optional[str] = None,
-        split_dataset: Optional[bool] = True,
-        rename: Optional[Dict] = None,
-        **kwargs,
-    ) -> List[str]:
-        r"""HYDROMT CORE METHOD: Add data variable(s) to maps object by reclassifying the data in ``raster_fn`` based on ``reclass_table_fn``.
-
-        This is done by reclassifying the data in
-        ``raster_fn`` based on ``reclass_table_fn``.
-
-        Adds model layers:
-
-        * **reclass_variables** maps: reclassified raster data
-
-        Parameters
-        ----------
-        raster_fn: str, Path, xr.DataArray
-            Data catalog key, path to raster file or raster xarray data object.
-            Should be a DataArray. Else use `variable` argument for selection.
-        reclass_table_fn: str, Path, pd.DataFrame
-            Data catalog key, path to tabular data file or tabular pandas dataframe
-            object for the reclassification table of `raster_fn`.
-        reclass_variables: list
-            List of reclass_variables from reclass_table_fn table to add to maps. Index
-            column should match values in `raster_fn`.
-        variable: str, optional
-            Name of raster dataset variable to use. This is only required when reading
-            datasets with multiple variables. By default None.
-        fill_method : str, optional
-            If specified, fills nodata values in `raster_fn` using fill_nodata method
-            before reclassifying. Available methods are {'linear', 'nearest',
-            'cubic', 'rio_idw'}.
-        reproject_method: str, optional
-            See rasterio.warp.reproject for existing methods, by default the data is
-            not reprojected (None).
-        name: str, optional
-            Name of new maps variable, only in case split_dataset=False.
-        split_dataset: bool, optional
-            If data is a xarray.Dataset split it into several xarray.DataArrays.
-        rename: dict, optional
-            Dictionary to rename variable names in reclass_variables before adding to
-            grid {'name_in_reclass_table': 'name_in_grid'}. By default empty.
-        \**kwargs:
-            Additional keyword arguments that are passed to the
-            `data_catalog.get_rasterdataset` function.
-
-        Returns
-        -------
-        list
-            Names of added model map layers
-        """  # noqa: E501
-        rename = rename or {}
-        self.logger.info(
-            f"Preparing map data by reclassifying the data in {raster_fn} based"
-            f" on {reclass_table_fn}"
-        )
-        # Read raster data and remapping table
-        da = self.data_catalog.get_rasterdataset(
-            raster_fn, geom=self.region, buffer=2, variables=variable, **kwargs
-        )
-        if not isinstance(da, xr.DataArray):
-            raise ValueError(
-                f"raster_fn {raster_fn} should be a single variable. "
-                "Please select one using the 'variable' argument"
-            )
-        df_vars = self.data_catalog.get_dataframe(
-            reclass_table_fn, variables=reclass_variables
-        )
-        # Fill nodata
-        if fill_method is not None:
-            da = da.raster.interpolate_na(method=fill_method)
-        # Mapping function
-        ds_vars = da.raster.reclassify(reclass_table=df_vars, method="exact")
-        # Reprojection
-        if ds_vars.rio.crs != self.crs and reproject_method is not None:
-            ds_vars = ds_vars.raster.reproject(dst_crs=self.crs)
-        # Add to maps
-        self.set_maps(ds_vars.rename(rename), name=name, split_dataset=split_dataset)
-
-        return list(ds_vars.data_vars.keys())
 
     @property
     def crs(self) -> CRS:
